use react-router-dom module for routing and rendering component
