+++ practice/4_ReactJS/1_react_movie_app/.eslintcache
@@ -1,1 +1,1 @@
-[{"/var/www/html/woosuk/HTML-CSS/practice/4_ReactJS/1_react_movie_app/src/index.js":"1","/var/www/html/woosuk/HTML-CSS/practice/4_ReactJS/1_react_movie_app/src/App.js":"2","/var/www/html/woosuk/HTML-CSS/practice/4_ReactJS/1_react_movie_app/src/Movie.js":"3"},{"size":197,"mtime":1610090594742,"results":"4","hashOfConfig":"5"},{"size":1161,"mtime":1610346949258,"results":"6","hashOfConfig":"5"},{"size":1007,"mtime":1610349292842,"results":"7","hashOfConfig":"5"},{"filePath":"8","messages":"9","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"10"},"146pio4",{"filePath":"11","messages":"12","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"13","messages":"14","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"/var/www/html/woosuk/HTML-CSS/practice/4_ReactJS/1_react_movie_app/src/index.js",[],["15","16"],"/var/www/html/woosuk/HTML-CSS/practice/4_ReactJS/1_react_movie_app/src/App.js",[],"/var/www/html/woosuk/HTML-CSS/practice/4_ReactJS/1_react_movie_app/src/Movie.js",[],{"ruleId":"17","replacedBy":"18"},{"ruleId":"19","replacedBy":"20"},"no-native-reassign",["21"],"no-negated-in-lhs",["22"],"no-global-assign","no-unsafe-negation"]
+[{"/var/www/html/woosuk/HTML-CSS/practice/4_ReactJS/1_react_movie_app/src/index.js":"1","/var/www/html/woosuk/HTML-CSS/practice/4_ReactJS/1_react_movie_app/src/App.js":"2","/var/www/html/woosuk/HTML-CSS/practice/4_ReactJS/1_react_movie_app/src/components/Movie.js":"3","/var/www/html/woosuk/HTML-CSS/practice/4_ReactJS/1_react_movie_app/src/routes/About.js":"4","/var/www/html/woosuk/HTML-CSS/practice/4_ReactJS/1_react_movie_app/src/routes/Home.js":"5","/var/www/html/woosuk/HTML-CSS/practice/4_ReactJS/1_react_movie_app/src/components/Navigation.js":"6"},{"size":197,"mtime":1610702046522,"results":"7","hashOfConfig":"8"},{"size":437,"mtime":1611564778995,"results":"9","hashOfConfig":"8"},{"size":1007,"mtime":1611561316843,"results":"10","hashOfConfig":"8"},{"size":393,"mtime":1611565187539,"results":"11","hashOfConfig":"8"},{"size":1229,"mtime":1611562832231,"results":"12","hashOfConfig":"8"},{"size":285,"mtime":1611564976954,"results":"13","hashOfConfig":"8"},{"filePath":"14","messages":"15","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"16"},"146pio4",{"filePath":"17","messages":"18","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"16"},{"filePath":"19","messages":"20","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"16"},{"filePath":"21","messages":"22","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"23","messages":"24","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"16"},{"filePath":"25","messages":"26","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"/var/www/html/woosuk/HTML-CSS/practice/4_ReactJS/1_react_movie_app/src/index.js",[],["27","28"],"/var/www/html/woosuk/HTML-CSS/practice/4_ReactJS/1_react_movie_app/src/App.js",[],"/var/www/html/woosuk/HTML-CSS/practice/4_ReactJS/1_react_movie_app/src/components/Movie.js",[],"/var/www/html/woosuk/HTML-CSS/practice/4_ReactJS/1_react_movie_app/src/routes/About.js",[],"/var/www/html/woosuk/HTML-CSS/practice/4_ReactJS/1_react_movie_app/src/routes/Home.js",[],"/var/www/html/woosuk/HTML-CSS/practice/4_ReactJS/1_react_movie_app/src/components/Navigation.js",[],{"ruleId":"29","replacedBy":"30"},{"ruleId":"31","replacedBy":"32"},"no-native-reassign",["33"],"no-negated-in-lhs",["34"],"no-global-assign","no-unsafe-negation"]
