+++ practice/4_ReactJS/1_react_movie_app/src/App.css
@@ -10,31 +10,3 @@ body {
   background-color: #eff3f7;
   height: 100%;
 }
-
-html,
-body,
-#root,
-.container {
-  height: 100%;
-  display: flex;
-  justify-content: center;
-}
-
-.loader {
-  width: 100%;
-  height: 100%;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  font-weight: 600;
-}
-
-.movies {
-  display: flex;
-  justify-content: space-between;
-  align-items: flex-start;
-  flex-wrap: wrap;
-  padding: 50px;
-  padding-top: 70px;
-  width: 80%;
-}
