+++ practice/4_ReactJS/1_react_movie_app/src/components/Movie.css
